requirements.txt
imbalanced-learn
jupyter
matplotlib
numpy
pandas
pre-commit
pyaml
PyYAML
ray
scikit-learn
scikit-optimize
scipy
seaborn
shap
simplejson
tune-sklearn
xgboost